billing-backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Allen Chen <allen.chen@hiero7.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
djangorestframework = "^3.14.0"
drf-jwt = "^1.19.2"
django = "^4.1.6"
mysqlclient = "^2.1.1"
django-redis = "^5.2.0"
redis = "^4.5.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
